fix(lib/shopt): turn on "extglob" by default
We thought that we already have `extglob` turned on, but it was just
turned on by bash-completion that we loaded from "lib/bourne-shell".
When the sytem does not have bash-completion, the extglob settings are
not enabled and thus causes an issue with constructs using extglob.
We already turn on many glob features in "lib/shopt", and there does
not seem to be reason not to turn on "extglob".  In this patch, we
enable it by default.