fix(themes/developer): avoid mixing stdout
This attempted to run the commands to extract information in parallel,
but this may result in interleaved information in some operating
system (depending on the way how the C library performs buffering and
how the syscall write(2) issued by multiple processes would interfere
with one another).

There are other parts we can optimize, so we would compensate the
overhead with the optimization of other parts.