fix(plugins/nvm): Correctly handle `.nvmrc` in a parent directory.
When checking the version in the resolved `.nvmrc` file found, the
change made in d2163a6 tried to read the file a second time, but did not
use the resolved `$nvm_path` but rather tried to simply read it from the
current directory. This fails if the file was resolved to a parent
directory.

This change updates the code to reuse the version value which was read
from the file two lines prior.