fix(tools/check_for_upgrade): fix broken check of uprade.lock (revert bdd2220)
This reverts commit bdd222055e7adab7a8b8185848d9706f4c63a6cd.

In commit bdd2220, the option -p was added to make sure to crete the
subdirectory "$OSH/log".  However, with the -p option, the command
mkdir succeeds even when a directory already exists, which conflicts
with the present purpose of crearing the lock directory only when
there is no existing one.

As long as we have log/.gitkeep in the repository, the "log" directory
is supposed to exist, so we do not need to make sure its existence.
We simply revert the affecting commit.